Revert "THRIFT-2045: C++ compiler warning and build cleanup"
This reverts commit 282e440c6de219b7b8f32b01cc7eb599f534f33f.

This change had unintended autoconf side effects.

+++ compiler/cpp/src/parse/t_field.h
@@ -155,7 +155,7 @@ class t_field : public t_doc {
  * not it was specified by the user or automatically chosen.
  */
 struct t_field_id {
-  int32_t value;
+  int64_t value;
   bool auto_assigned;
 };
